I maked code review in backend
