+++ routes/web.php
@@ -12,22 +12,15 @@ use Illuminate\Support\Facades\Route;
 | contains the "web" middleware group. Now create something great!
 |
 */
-//Route::get('/reactHome', 'HomeController@reactHome');
 Route::get('/', 'HomeController@reactHome');
 Route::get('/reactRoom/{id}', 'HomeController@reactRoom');
-//Route::get('/', 'HomeController@index');
-//Route::get('/room/{id}', 'HomeController@room')->name('room');
-Route::get('/generateLettersArray/{id}', 'RoomController@generateLettersArray');
-Route::get('/generate', 'RoomController@generate');
-Route::get('/checkLogin', 'RoomController@checkLogin');
-Route::get('/checkLettersArray', 'RoomController@checkLettersArray');
-Route::get('/getPlayers', 'RoomController@getPlayers');
-//Route::get('/login', 'RoomController@login');
-Route::post('/login', 'RoomController@login');
-Route::post('/saveWords', 'RoomController@saveWords');
-Route::get('/resetRoom', 'RoomController@resetRoom');
+
+Route::get('/checkLogin', 'CheckController@checkLogin');
 Route::get('/checkDictionary', 'CheckController@checkDictionary');
 
-//Auth::routes();
+Route::get('/getPlayers', 'RoomController@getPlayers');
+Route::get('/generateLettersArray', 'RoomController@generateLettersArray');
+Route::get('/resetRoom', 'RoomController@resetRoom');
 
-//Route::get('/home', 'HomeController@index')->name('home');
+Route::post('/login', 'RoomController@login');
+Route::post('/saveWords', 'RoomController@saveWords');
